Fixed some bugs introduced during operation overhaul

+++ net.dougqh.jak.jvm.core/src/net/dougqh/jak/jvm/operations/lstore_0.java
@@ -36,7 +36,7 @@ public final class lstore_0 extends StoreOperation {
 	
 	@Override
 	public final Type type() {
-		return int.class;
+		return long.class;
 	}
 	
 	@Override
